+++ client/v2/autocli/flag/builder.go
@@ -16,7 +16,6 @@ import (
 
 	autocliv1 "cosmossdk.io/api/cosmos/autocli/v1"
 	msgv1 "cosmossdk.io/api/cosmos/msg/v1"
-	"cosmossdk.io/client/v2/autocli/keyring"
 	"cosmossdk.io/client/v2/internal/flags"
 	"cosmossdk.io/client/v2/internal/util"
 	"cosmossdk.io/core/address"
@@ -47,9 +46,6 @@ type Builder struct {
 
 	messageFlagTypes map[protoreflect.FullName]Type
 	scalarFlagTypes  map[string]Type
-
-	// Keyring is the keyring to use for client/v2.
-	Keyring keyring.Keyring
 
 	// Address Codecs are the address codecs to use for client/v2.
 	AddressCodec          address.Codec
@@ -90,10 +86,6 @@ func (b *Builder) ValidateAndComplete() error {
 		return errors.New("consensus address codec is required in flag builder")
 	}
 
-	if b.Keyring == nil {
-		b.Keyring = keyring.NoKeyring{}
-	}
-
 	if b.TypeResolver == nil {
 		return errors.New("type resolver is required in flag builder")
 	}
@@ -117,6 +109,30 @@ func (b *Builder) DefineScalarFlagType(scalarName string, flagType Type) {
 	b.scalarFlagTypes[scalarName] = flagType
 }
 
+// GetMinimumArgs returns the minimum number of positional arguments required for the command.
+func (b *Builder) GetMinimumArgs(commandOptions *autocliv1.RpcCommandOptions) (cobra.PositionalArgs, error) {
+	positionalArgsLen := len(commandOptions.PositionalArgs)
+	for i, arg := range commandOptions.PositionalArgs {
+		if arg.Varargs {
+			if i != positionalArgsLen-1 {
+				return nil, fmt.Errorf("varargs positional argument %s must be the last argument", arg.ProtoField)
+			}
+
+			return cobra.MinimumNArgs(positionalArgsLen - 1), nil
+		}
+
+		if arg.Optional {
+			if i != positionalArgsLen-1 {
+				return nil, fmt.Errorf("optional positional argument %s must be the last argument", arg.ProtoField)
+			}
+
+			return cobra.RangeArgs(positionalArgsLen-1, positionalArgsLen), nil
+		}
+	}
+
+	return cobra.ExactArgs(positionalArgsLen), nil
+}
+
 // AddMessageFlags adds flags for each field in the message to the flag set.
 func (b *Builder) AddMessageFlags(ctx context.Context, flagSet *pflag.FlagSet, messageType protoreflect.MessageType, commandOptions *autocliv1.RpcCommandOptions) (*MessageBinder, error) {
 	return b.addMessageFlags(ctx, flagSet, messageType, commandOptions, namingOptions{})
@@ -135,7 +151,7 @@ func (b *Builder) addMessageFlags(ctx context.Context, flagSet *pflag.FlagSet, m
 
 	isPositional := map[string]bool{}
 
-	lengthPositionalArgsOptions := len(commandOptions.PositionalArgs)
+	positionalArgsLen := len(commandOptions.PositionalArgs)
 	for i, arg := range commandOptions.PositionalArgs {
 		isPositional[arg.ProtoField] = true
 
@@ -147,17 +163,12 @@ func (b *Builder) addMessageFlags(ctx context.Context, flagSet *pflag.FlagSet, m
 			}
 		}
 
-		field := fields.ByName(protoreflect.Name(arg.ProtoField))
-		if field == nil {
-			return nil, fmt.Errorf("can't find field %s on %s", arg.ProtoField, messageType.Descriptor().FullName())
-		}
-
 		if arg.Optional && arg.Varargs {
 			return nil, fmt.Errorf("positional argument %s can't be both optional and varargs", arg.ProtoField)
 		}
 
 		if arg.Varargs {
-			if i != lengthPositionalArgsOptions-1 {
+			if i != positionalArgsLen-1 {
 				return nil, fmt.Errorf("varargs positional argument %s must be the last argument", arg.ProtoField)
 			}
 
@@ -165,11 +176,16 @@ func (b *Builder) addMessageFlags(ctx context.Context, flagSet *pflag.FlagSet, m
 		}
 
 		if arg.Optional {
-			if i != lengthPositionalArgsOptions-1 {
+			if i != positionalArgsLen-1 {
 				return nil, fmt.Errorf("optional positional argument %s must be the last argument", arg.ProtoField)
 			}
 
 			messageBinder.hasOptional = true
+		}
+
+		field := fields.ByName(protoreflect.Name(arg.ProtoField))
+		if field == nil {
+			return nil, fmt.Errorf("can't find field %s on %s", arg.ProtoField, messageType.Descriptor().FullName())
 		}
 
 		_, hasValue, err := b.addFieldFlag(
@@ -189,15 +205,16 @@ func (b *Builder) addMessageFlags(ctx context.Context, flagSet *pflag.FlagSet, m
 		})
 	}
 
-	if messageBinder.hasVarargs {
-		messageBinder.CobraArgs = cobra.MinimumNArgs(lengthPositionalArgsOptions - 1)
-		messageBinder.mandatoryArgUntil = lengthPositionalArgsOptions - 1
-	} else if messageBinder.hasOptional {
-		messageBinder.CobraArgs = cobra.RangeArgs(lengthPositionalArgsOptions-1, lengthPositionalArgsOptions)
-		messageBinder.mandatoryArgUntil = lengthPositionalArgsOptions - 1
-	} else {
-		messageBinder.CobraArgs = cobra.ExactArgs(lengthPositionalArgsOptions)
-		messageBinder.mandatoryArgUntil = lengthPositionalArgsOptions
+	switch {
+	case messageBinder.hasVarargs:
+		messageBinder.CobraArgs = cobra.MinimumNArgs(positionalArgsLen - 1)
+		messageBinder.mandatoryArgUntil = positionalArgsLen - 1
+	case messageBinder.hasOptional:
+		messageBinder.CobraArgs = cobra.RangeArgs(positionalArgsLen-1, positionalArgsLen)
+		messageBinder.mandatoryArgUntil = positionalArgsLen - 1
+	default:
+		messageBinder.CobraArgs = cobra.ExactArgs(positionalArgsLen)
+		messageBinder.mandatoryArgUntil = positionalArgsLen
 	}
 
 	// validate flag options