+++ client/v2/autocli/flag/interface.go
@@ -9,8 +9,7 @@ import (
 
 // Type specifies a custom flag type.
 type Type interface {
-	// NewValue returns a new pflag.Value which must also implement either
-	// SimpleValue or ListValue.
+	// NewValue returns a new pflag.Value which must also implement either SimpleValue or ListValue.
 	NewValue(context.Context, *Builder) Value
 
 	// DefaultValue is the default value for this type.